identity/pyproject.toml
[tool.poetry]
name = "clara-identity"
version = "0.1.0"
description = "OAuth identity service for MyPalClara"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11,<3.14"
fastapi = "^0.115.0"
uvicorn = "^0.38.0"
sqlalchemy = "^2.0"
psycopg2-binary = "^2.9.9"
pyjwt = "^2.8.0"
httpx = "^0.28.0"
python-dotenv = "^1.0.0"
pydantic = "^2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-asyncio = "^0.24.0"
ruff = "^0.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
